+++ agape/src/widgets/container.rs
@@ -56,6 +56,7 @@ impl<W: Widget> Widget for Container<W> {
             &self.style.background_color.clone(),
             size,
             position,
+            self.style.corner_radius,
             self.style.border.clone(),
         );
         self.child.render(pixmap, renderer, layout);