+++ agape/src/widgets/hstack.rs
@@ -116,6 +116,7 @@ impl Widget for HStack {
             &self.style.background_color.clone(),
             size,
             position,
+            self.style.corner_radius,
             self.style.border.clone(),
         );
         // TODO: test this